+++ test/clitest.hs
@@ -35,11 +35,11 @@ testEnv = Env { config = defaultConfig {
   , verb = 1
   } }
 
-runForgeTest :: FilePath -> [String] -> IO (ExitCode, String, String)
-runForgeTest testFile extraOptions = do
+runForge :: FilePath -> [String] -> IO (ExitCode, String, String)
+runForge testFile extraOptions = do
   withSystemTempDirectory "dapp-test" $ \root -> do
     let projectType = Foundry
-    ret <- runEnv testEnv $ compile projectType root testFile
+    ret <- runEnv testEnv $ compileWithForge root testFile
     case ret of
       Left e -> do
         putStrLn e
@@ -189,34 +189,34 @@ main = do
 
       -- both should fail with address 0x1234 -- a SPECIFIC address, not a ranomly generated one
       it "keccak-assumptions-setup" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-setup.sol" []
+        (exitCode, stdout, stderr) <- runForge "test/contracts/fail/keccak-preimage-setup.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[validated]"
         stdout `shouldContain` "0x0000000000000000000000000000000000001234"
         exitCode `shouldBe` (ExitFailure 1)
       it "keccak-assumptions-constructor" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-preimage-constructor.sol" []
+        (exitCode, stdout, stderr) <- runForge "test/contracts/fail/keccak-preimage-constructor.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[validated]"
         stdout `shouldContain` "0x0000000000000000000000000000000000001234"
         exitCode `shouldBe` (ExitFailure 1)
       it "keccak-wrong-preimage" $ do
-        (exitCode, stdout, stderr) <- runForgeTest "test/contracts/fail/keccak-wrong-preimage.sol" []
+        (exitCode, stdout, stderr) <- runForge "test/contracts/fail/keccak-wrong-preimage.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[not reproducible]"
         exitCode `shouldBe` (ExitFailure 1)
       it "only-deployed-contracts" $ do
-        (_, stdout, stderr) <- runForgeTest "test/contracts/pass/only-deployed-contracts.sol" ["--only-deployed"]
+        (_, stdout, stderr) <- runForge "test/contracts/pass/only-deployed-contracts.sol" ["--only-deployed"]
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[PASS]"
       it "only-deployed-contracts-force-addr" $ do
-        (_, stdout, stderr) <- runForgeTest "test/contracts/pass/only-deployed-force-addr.sol" ["--only-deployed"]
+        (_, stdout, stderr) <- runForge "test/contracts/pass/only-deployed-force-addr.sol" ["--only-deployed"]
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[FAIL]"
         stdout `shouldContain` "[validated]"
         stderr `shouldNotContain` "not reproducible"
       it "should-fail" $ do
-        (_, stdout, stderr) <- runForgeTest "test/contracts/fail/should-fail.sol" []
+        (_, stdout, stderr) <- runForge "test/contracts/fail/should-fail.sol" []
         stderr `shouldNotContain` "CallStack"
         stdout `shouldContain` "[FAIL]"
         stdout `shouldContain` "[validated]"
@@ -242,7 +242,7 @@ main = do
         shouldBe fileExists True
         removeFile filename
       it "rpc-mock" $ do
-        (_, stdout, stderr) <- runForgeTest "test/contracts/fail/rpc-test.sol"
+        (_, stdout, stderr) <- runForge "test/contracts/fail/rpc-test.sol"
           ["--rpc", "http://mock.mock", "--prefix", "test_attack_symbolic"
           , "--number", "10307563", "--mock-file", "test/contracts/fail/rpc-test-mock.json"]
         stdout `shouldContain` "[FAIL]"